refactor metadata module & test imports
