+++ Makefile
@@ -2,7 +2,7 @@ pytest:
 	python -m pytest --log-cli-level info -p no:warnings -v ./tests
 
 format:
-	python -m black -S --line-length 79 --experimental-string-processing ./
+	python -m black -S --line-length 79 --preview ./
 	isort ./
 
 type:
