+++ README.md
@@ -41,7 +41,4 @@ The following commands are to be run via the terminal, inside your project root 
 
 ## Make commands
 
-We have some make commands to make things run better
-
-1. **make ci**: Runs code formatting, lint checking, type checking and pytests. You can also run them individually with `make format`, `make lint`, `make type`, and `make pytests` respectively.
-2. **make twitter-etl** and **make reddit-etl**: Runs the twitter and reddit etl with standard settings. Use `python ./socialetl/main.py --help` to determine the parameters and run them using commands like `python ./socialetl/main.py --etl reddit --tx no_tx --log info --reset-db`, etc.
+We have some make commands to make things run better, please refer to the [Makefile](./Makefile) to see them.
